Making graffle2pdf a standalone AppleScript tool
This version combining to some extent graffle.scpt and graffle.sh to an
AppleScript graffle2pdf that can be used as a stand-alone command-line
tool.  graffle2pdf, unlike the old graffle.scpt is a plain-text
AppleScript that properly handles relative/absolute file names of the
supplied arguments.

A small regression is that this script requires OmniGraffle Professional.